rework types, using @types/emscripten such as it is
