+++ packages/contents/src/index.ts
@@ -5,3 +5,4 @@ export * from './contents';
 export * from './drivefs';
 export * from './tokens';
 export * from './broadcast';
+export * from './emscripten';
